+++ include/streaming/dash_manifest_parser.h
@@ -8,9 +8,13 @@ namespace vp::streaming {
 struct DashRepresentation {
     std::string id;
     int bandwidth{0};
+    std::string base_url;
+    std::string initialization_url;
+    std::vector<std::string> segment_urls;
 };
 
 struct DashManifest {
+    std::string base_url;
     std::vector<DashRepresentation> representations;
 };